Move state to redux
