+++ src/redux/store.ts
@@ -1,0 +1,7 @@
+import {createStore, combineReducers, applyMiddleware} from 'redux';
+import thunk from 'redux-thunk';
+import userReducer from './reducers';
+
+const rootReducer = combineReducers({userReducer});
+
+export const Store = createStore(rootReducer);
